feat: add enum class and lotto result logic
