+++ src/main/kotlin/lotto/Application.kt
@@ -3,17 +3,7 @@ package lotto
 import camp.nextstep.edu.missionutils.Console
 import camp.nextstep.edu.missionutils.Randoms
 
-enum class LottoError(val message: String) {
-    INVALID_AMOUNT("[ERROR] 유효한 금액을 입력해주세요."),
-    INVALID_UNIT("[ERROR] 로또 금액은 1000원 단위로 입력해야 합니다."),
-    INVALID_NUM("[ERROR] 1~45 사이의 중복되지 않는 6개의 숫자를 입력해 주세요."),
-    INVALID_SPLIT("[ERROR] 유효한 구분자와 숫자를 사용해 주세요."),
-    INVALID_DUPLICATE("[ERROR] 보너스 번호는 당첨 번호와 중복될 수 없습니다.");
 
-    fun throwException(): IllegalArgumentException {
-        return IllegalArgumentException(this.message)
-    }
-}
 
 fun main() {
     val amount = readAndValidateAmount()
@@ -23,6 +13,7 @@ fun main() {
 
     val winningNumbers = readLottoNumber()
     val bonusNumber = readBonusNumber(winningNumbers)
+
 }
 
 fun readAndValidateAmount(): Int {
@@ -63,7 +54,7 @@ fun readLottoNumber(): List<Int> {
     while (true) {
         try {
             val lottoNum = promptUserForLotto()
-            Validation.validateLottoNumbers(lottoNum) // 번호 유효성 검증
+            Validation.validateLottoNumbers(lottoNum)
             return lottoNum
         } catch (e: IllegalArgumentException) {
             println(e.message)
