+++ src/main/kotlin/lotto/Lotto.kt
@@ -5,5 +5,11 @@ class Lotto(private val numbers: List<Int>) {
         require(numbers.size == 6) { "[ERROR] 로또 번호는 6개여야 합니다." }
     }
 
-    // TODO: 추가 기능 구현
+    fun getMatchCount(winningNumbers: List<Int>): Int {
+        return numbers.intersect(winningNumbers.toSet()).size
+    }
+
+    fun containsNumber(number: Int): Boolean {
+        return numbers.contains(number)
+    }
 }
