+++ src/main/kotlin/lotto/LottoResult.kt
@@ -1,0 +1,23 @@
+package lotto
+
+enum class LottoResult(val matchCount: Int, val prize: Long) {
+    FIRST(6, 2_000_000_000),
+    SECOND(5, 30_000_000),
+    THIRD(5, 1_500_000),
+    FOURTH(4, 50_000),
+    FIFTH(3, 5_000),
+    NONE(0, 0);
+
+    companion object {
+        fun getResult(matchCount: Int, matchBonus: Boolean): LottoResult {
+            return when {
+                matchCount == 6 -> FIRST
+                matchCount == 5 && matchBonus -> SECOND
+                matchCount == 5 -> THIRD
+                matchCount == 4 -> FOURTH
+                matchCount == 3 -> FIFTH
+                else -> NONE
+            }
+        }
+    }
+}
